pyproject.toml
[project]
name = "nfvcl"
version = "0.4.0"
description = ""
authors = [{ name = "TNT LAB", email = "staff@tnt-lab.unige.it" }]
requires-python = "~=3.14"
readme = "README.md"
#license = "n\b"
maintainers = [
    { name = "Paolo Bono", email = "paolob@tnt-lab.unige.it" },
    { name = "Alderico Gallo", email = "aldericog@tnt-lab.unige.it" },
    { name = "Davide Freggiaro", email = "davidef@tnt-lab.unige.it" },
]
dependencies = [
    "fastapi>=0.119.1",
    "file_read_backwards>=3.2.0",
    "ruamel.yaml>=0.18.16",
    "jinja2-ansible-filters>=1.3.2,<2",
    "redis>=7.0.0",
    "ruamel.yaml.jinja2>=0.2.7,<0.3",
    "kubernetes==32.0.1",
    "uvicorn>=0.38.0",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pyOpenSSL==25.3.0",
    "pymongo>=4.15.3",
    "paramiko>=4.0.0",
    "requests>=2.32.5",
    "openstacksdk>=4.7.1",
    "coloredlogs>=15.0.1,<16",
    "colorama>=0.4.6,<0.5",
    "verboselogs~=1.7",
    "ansible-runner>=2.4.2",
    "ansible>=12.1.0",
    "httpx[http2]>=0.28.1",
    "jinja2>=3.1.6,<4",
    "dependency-injector>=4.48.2",
    "decohints>=1.0.9,<2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20,<0.0.21",
    "blinker>=1.9.0,<2",
    "proxmoxer>=2.2.0,<3",
    "grafana-client>=5.0.1",
    "pydantic-extra-types>=2.10.6",
    "pyhelm3",
    "pip"
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-dependency>=0.6.0,<0.7",
    "mypy>=1.18.2",
]

[tool.uv]

[tool.uv.sources]
pyhelm3 = { git = "https://github.com/s2n-cnit/pyhelm3.git", rev = "5c25193" }

[tool.hatch.build.targets.sdist]
include = [
    "src/nfvcl_core_models",
    "src/nfvcl_core",
    "src/nfvcl_rest",
    "src/nfvcl_horse",
    "src/nfvcl_models",
    "src/nfvcl",
]

[tool.hatch.build.targets.wheel]
include = [
    "src/nfvcl_core_models",
    "src/nfvcl_core",
    "src/nfvcl_rest",
    "src/nfvcl_horse",
    "src/nfvcl_models",
    "src/nfvcl",
]

[tool.hatch.build.targets.wheel.sources]
"src/nfvcl_core_models" = "nfvcl_core_models"
"src/nfvcl_core" = "nfvcl_core"
"src/nfvcl_rest" = "nfvcl_rest"
"src/nfvcl_horse" = "nfvcl_horse"
"src/nfvcl_models" = "nfvcl_models"
"src/nfvcl" = "nfvcl"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
